pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "species-identifier"
version = "0.1.0"
description = "A graph-based suite for identifying and counting chemical species from MD trajectories."
readme = "README.md"
requires-python = ">=3.9"
dependencies =[
    "numpy>=1.20.0",
    "ase>=3.22.0",
    "pyyaml>=5.4.0"
]

[project.scripts]
species-analyze = "species_identifier.analyzer:main"